p1_etiquette/requirements.txt
pillow
pdf2image
numpy
matplotlib
